feat(bcgov/footer): update story
update to mdx format

#294
