+++ packages/bcgov-theme/stories/footer/Footer.stories.mdx
@@ -1,0 +1,27 @@
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
+import Footer from '../../src/Footer';
+import * as stories from './Footer.stories.tsx';
+import { ArgsTable } from '@storybook/addon-docs/blocks';
+import { argTypes } from './args';
+
+<Meta title="Components/Footer" component={Footer} argTypes={argTypes} />
+
+# Footer
+
+Footers can be used to highlight content within a styled box. Any children passed to the Footer will be rendered into the main body.
+
+export const Template = (args) => <Footer {...args} />
+
+<Canvas>
+  <Story story={stories.Default} name="Default" />
+</Canvas>
+
+## Arguments
+
+The footer does not require any props. To render content, pass links as children in an unordered list. See the code block above for an example.
+
+## HTML only
+
+<Canvas>
+  <Story story={stories.HTML} name="html" />
+</Canvas>
