+++ packages/bcgov-theme/stories/footer/args.ts
@@ -1,0 +1,1 @@
+export const argTypes = {};
